+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 2nd November 2020
 
 - Initial commit
+- Add `ArgDetail` model
+	- And tests